refactor: split synthetic anomaly specs
